Time: 15 ms (18.66%) | Memory: 21 MB (83.90%) - LeetSync

+++ 169-majority-element/majority-element.py
@@ -15,6 +15,6 @@ class Solution:
         for i in range(n):
             if(nums[i] == element):
                 count1 += 1
-        if(count1 > int(n/2)):
+        if(count1 > n/2):
             return element
         return -1
